Extract getDefaultValues
